extract Creator relationship to concern
