+++ lib/engines/content_object_store/app/models/concerns/content_object_store/has_authors.rb
@@ -1,6 +1,7 @@
 module ContentObjectStore
   module HasAuthors
     extend ActiveSupport::Concern
+    include ContentObjectStore::HasCreator
 
     included do
       has_many :content_block_edition_authors, dependent: :destroy
